adds client route and link for new GameForm
